clean code
# message auto-generated for no-merge-commit merge:
merge yuwangang into master

clean code

Created-by: msprofDeveloper

Match-id-21d1fe398742c090bc591e5c44147090054a3197

+++ collector/dvvp/job_wrapper/src/prof_timer.cpp
@@ -1014,8 +1014,8 @@ bool NetDevStatsHandler::GetDcmiCardDevId(uint32_t devId, int &dcmiCardId, int &
         MSPROF_LOGE("NetDevStatsHandler get device num in card failed, deviceNum is invalid");
         return false;
     }
-    dcmiCardId = cardList[devId / deviceNum];
-    dcmiDevId = devId % deviceNum;
+    dcmiCardId = cardList[static_cast<long>(devId) / deviceNum];
+    dcmiDevId = static_cast<long>(devId) % deviceNum;
     return true;
 }
